% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_contri_pie_to_tree}
\alias{add_contri_pie_to_tree}
\title{add piecharts to a pre-plotted tree}
\usage{
add_contri_pie_to_tree(plot, tree, signature, cex)
}
\arguments{
\item{plot}{a plot produced by ggtree to add the piecharts to}

\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{signature}{the signature for which to plot the piecharts}

\item{cex}{the size of the piechart}

\item{sig_matrix}{the signature}
}
\value{
the ggplot with the piecharts added
}
\description{
add piecharts to a pre-plotted tree
}
