% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_contribution}
\alias{add_contribution}
\title{add a contribution matrix to a tree}
\usage{
add_contribution(
  tree,
  contribution = NULL,
  signatures = NULL,
  mut_matrix = NULL,
  ...
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{contribution}{a contribution matrix with signatures in the rows and samples in the columns}

\item{signatures}{a matrix with signature names in the columns and mutations in the rows}

\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}

\item{...}{parameters passed on to \code{\link[MutationalPatterns]{fit_to_signatures_strict}}}
}
\value{
a tree with a new data column for each signature in "contribution"
}
\description{
add a contribution matrix to a tree
}
