% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_info_to_tree}
\alias{add_info_to_tree}
\title{add additional branch/node informationi to the tree}
\usage{
add_info_to_tree(tree, by = "branch_id", ...)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{by}{the column in \code{tree@data} to join by}

\item{...}{info to add to the tree.
Either data.frames/tibbles with a column specified in \code{by}, containing the branch ids (as present in \code{tree@data[ ,by]})
or a numeric/factor/character named with the branch ids found in \code{tree@data[ ,by]}}
}
\value{
a tree with the additional stored information
}
\description{
add additional branch/node informationi to the tree
}
