% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_tree_drivers}
\alias{add_tree_drivers}
\title{add information on (potential) driver mutations to a tree}
\usage{
add_tree_drivers(
  tree,
  branch_vcf_list,
  add_mod_high = TRUE,
  add_cosmic = FALSE,
  blacklist = NULL,
  include_list = NULL
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{branch_vcf_list}{a list of \code{CollapsedVCFs}, one for each branch}

\item{add_mod_high}{a boolean: whether or not to include MODERATE/HIGH impact mutations}

\item{add_cosmic}{a boolean: whether or not to include mutations with a COSMIC ID (i.e., at least 1 sample registered in the COSMIC database)}

\item{blacklist}{a (regex) string which will be excluded from the driver list, e.g. "^MUC.*" to remove mutation in MUC genes}

\item{include_list}{a vector of gene names to include in the driver list (ignoring all other genes)}
}
\value{
a tree with a new column "drivers" containting potential cancer drives
}
\description{
add information on (potential) driver mutations to a tree
}
