% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{change_tip_labels}
\alias{change_tip_labels}
\title{change the tip.labels of your tree}
\usage{
change_tip_labels(tree, id_change_tb)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{id_change_tb}{a data.frame with 2 columns, the first the current tip.labels (to change), and the second, the new corresponding IDs}
}
\value{
a tree, with a \code{branch_id} column for branch identification
}
\description{
change the tip.labels of your tree
}
