% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_0_mm}
\alias{check_0_mm}
\title{remove columns with 0 (or another defined number of) mutations from a mutational matrix}
\usage{
check_0_mm(mut_matrix, remove_min = 0)
}
\arguments{
\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}

\item{remove_min}{columns with equal or fewer mutations that this number will be removed}
}
\description{
remove columns with 0 (or another defined number of) mutations from a mutational matrix
}
