% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_reconstructed_cosine}
\alias{check_reconstructed_cosine}
\title{plotting whether a reconstructed profile (after signature refitting) is similar enough to the original profiel}
\usage{
check_reconstructed_cosine(contribution, mut_matrix, signatures, tree)
}
\arguments{
\item{contribution}{a contribution matrix with signatures in the rows and samples in the columns}

\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}

\item{signatures}{a matrix with signature names in the columns and mutations in the rows}

\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}
}
\value{
a plot comparing the number of mutation per profile to the cosine similarity between the reconstructed vs the original profile
}
\description{
plotting whether a reconstructed profile (after signature refitting) is similar enough to the original profiel
}
