% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_vcf_to_granges}
\alias{convert_vcf_to_granges}
\title{convert a list of VCFs to a GRangesList}
\usage{
convert_vcf_to_granges(branch_vcf_list, ref_genome, incl_chroms = 1:24)
}
\arguments{
\item{branch_vcf_list}{a list of \code{CollapsedVCFs} to convert to a \code{\link[GenomicRanges]{GRangesList}}}

\item{ref_genome}{a character string: the reference BSgenome}

\item{incl_chroms}{a numeric: which chromosomes to include from the ref_genome}
}
\value{
a GRangesList
}
\description{
convert a list of VCFs to a GRangesList
}
