% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_vcf_per_branch}
\alias{extract_vcf_per_branch}
\title{Title}
\usage{
extract_vcf_per_branch(
  tree,
  vcf,
  make_granges = FALSE,
  ref_genome = NA,
  incl_chroms = 1:24
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{vcf}{the VCF used to build the tree as a \code{CollapsedVCF}, i.e., imported by \code{\link[VariantAnnotation]{readVcf}}}

\item{make_granges}{a boolean: whether to output a list of \code{CollapsedVCFs} (\code{FALSE}) or a \code{\link[GenomicRanges]{GRangesList}} (\code{TRUE})}

\item{ref_genome}{a character string: the reference BSgenome}

\item{incl_chroms}{a numeric: which chromosomes to include from the ref_genome}
}
\value{
the mutations per branch, either as a list of \code{CollapsedVCFs} or a \code{\link[GenomicRanges]{GRangesList}}
}
\description{
Title
}
