% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_common_sample_string}
\alias{find_common_sample_string}
\title{find the longest common substring in the sample IDs}
\usage{
find_common_sample_string(tree)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}
}
\value{
the longest common substring in the sample IDs
}
\description{
find the longest common substring in the sample IDs
}
