% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_gene}
\alias{find_gene}
\title{Find mutations in certain genes in a list of}
\usage{
find_gene(branch_vcf_list, find_gene)
}
\arguments{
\item{branch_vcf_list}{a list of \code{CollapsedVCFs}, one for each branch}

\item{find_gene}{}
}
\value{
a string of matching mutations in the genes, named by the branch_id
}
\description{
Find mutations in certain genes in a list of
}
