% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_to_signatures_strict_tree}
\alias{fit_to_signatures_strict_tree}
\title{A wrapper that runs and filters the output of \code{\link[MutationalPatterns]{fit_to_signatures_strict}}}
\usage{
fit_to_signatures_strict_tree(
  mut_matrix,
  signatures,
  max_delta = 0.01,
  remove_min = 20
)
}
\arguments{
\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}

\item{signatures}{a matrix with signature names in the columns and mutations in the rows}

\item{max_delta}{see \code{\link[MutationalPatterns]{fit_to_signatures_strict}}}

\item{remove_min}{columns with equal or fewer mutations that this number will be removed}
}
\value{
a contribution matrix with signatures in the rows and samples in the columns
}
\description{
A wrapper that runs and filters the output of \code{\link[MutationalPatterns]{fit_to_signatures_strict}}
}
