% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_tree_contributions}
\alias{group_tree_contributions}
\title{add the contribution matrix columns from a group together}
\usage{
group_tree_contributions(
  tree,
  group,
  contribution = NULL,
  signatures = NULL,
  mut_matrix = NULL,
  ...
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{group}{a column in \code{tree@data} that specifies the group by which to summarize the contribution matrix}

\item{contribution}{a contribution matrix with signatures in the rows and samples in the columns}

\item{signatures}{a matrix with signature names in the columns and mutations in the rows}

\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}

\item{...}{parameters passed on to \code{\link[MutationalPatterns]{fit_to_signatures_strict}}}
}
\value{
a contribution matrix with one column per group, containing the sum of the mutation numbers in each group
}
\description{
add the contribution matrix columns from a group together
}
