% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_tree_mut_matrix}
\alias{group_tree_mut_matrix}
\title{add the mutational matrix columns from a group together}
\usage{
group_tree_mut_matrix(tree, group, mut_matrix)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{group}{a column in \code{tree@data} that specifies the group by which to summarize the mutational matrix}

\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}
}
\value{
a mutational matrix with one column per group, containing the sum of the mutation numbers in each group
}
\description{
add the mutational matrix columns from a group together
}
