% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_96_profile2}
\alias{plot_96_profile2}
\title{an adjustment on MutationalPatterns' plot_96_profile}
\usage{
plot_96_profile2(
  mut_matrix,
  colors = NA,
  ymax = NULL,
  condensed = FALSE,
  addLine = TRUE,
  small = FALSE,
  axisnormal = FALSE,
  normalize = TRUE,
  ann_size = 3,
  label_x = "A[T>C]A",
  ncol = 1
)
}
\value{
a ggplot
}
\description{
an adjustment on MutationalPatterns' plot_96_profile
}
