% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gg_tree}
\alias{plot_gg_tree}
\title{plot a tree}
\usage{
plot_gg_tree(
  tree,
  common_name = NA,
  add_x = TRUE,
  add_tip_label = "short_tip_lab",
  add_bootstrap = FALSE,
  add_branch_length = FALSE,
  add_branch_id = FALSE,
  branch_text_param = NULL,
  branch_text_size = 3,
  only_shared_branches = FALSE,
  plot_margin = unit(c(10, 100, 10, 10), "points"),
  branch_color_param = NULL,
  branch_colors = NA,
  add_title = TRUE,
  legend_pos = NULL
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{common_name}{a string to substract from the sample names and use to plot above the plot. (default: the longest common sample substring)}

\item{add_x}{a boolean: whether to plot an x-axis}

\item{add_tip_label}{a boolean: whether to plot the sample names at the tips of the leaves (i.e., end branches)}

\item{add_bootstrap}{a boolean: whether to plot the bootstrap information on the nodes}

\item{add_branch_length}{a boolean: whether to plot the branch length numbers}

\item{add_branch_id}{a boolean: whether to plot the branch IDs}

\item{branch_text_param}{a character: which additional text to plot at each branch.
Should be one of the column names of \code{tree@data}.}

\item{branch_text_size}{the text size of the 'branch_text_param' parameter}

\item{only_shared_branches}{a boolean: whether to only plot the shared branches}

\item{plot_margin}{the plot margin (default: \code{unit(c(10,100,10,10), 'points')})}

\item{branch_color_param}{a character: how to color each branch.
Should be one of the column names of \code{tree@data}.}

\item{branch_colors}{the colors used for coloring the branches}

\item{add_title}{a character with the title to plot. Or a boolean: whether to plot the 'common_name' as a title}

\item{legend_pos}{a specification of the legend position, e.g. "bottom". See \code{\link[ggplot2]{theme}}, "legend.position"}
}
\value{
a ggplot
}
\description{
plot a tree
}
