% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gg_tree_base}
\alias{plot_gg_tree_base}
\title{the plot_gg_tree with some default settings, like plotting the branch IDs}
\usage{
plot_gg_tree_base(
  tree,
  add_tip_label = "short_tip_lab",
  add_bootstrap = TRUE,
  add_branch_length = TRUE,
  add_branch_id = TRUE,
  ...
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{add_tip_label}{a boolean: whether to plot the sample names at the tips of the leaves (i.e., end branches)}

\item{add_bootstrap}{a boolean: whether to plot the bootstrap information on the nodes}

\item{add_branch_length}{a boolean: whether to plot the branch length numbers}

\item{add_branch_id}{a boolean: whether to plot the branch IDs}

\item{...}{parameters passed on to \code{\link{plot_gg_tree}}}
}
\value{
a ggplot
}
\description{
the plot_gg_tree with some default settings, like plotting the branch IDs
}
