% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tree_contribution}
\alias{plot_tree_contribution}
\title{plot the tree with the contribution of signature per branch}
\usage{
plot_tree_contribution(tree, signature, type = "pie", pie_size = 1, ...)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{signature}{the signature to plot. Should have been add before with \link{add_contribution}}

\item{type}{a character. "pie" to plot a pie chart, 'color' to color the branches, 'color_dot' to add a colored dot}

\item{pie_size}{the size of the piechart per branch if type = 'pie'. (default = 1)}

\item{...}{parameters to pass on to \link{plot_gg_tree}}
}
\value{
a ggplot
}
\description{
plot the tree with the contribution of signature per branch
}
