% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tree_contribution_bars}
\alias{plot_tree_contribution_bars}
\title{plot the tree with the contribution of signature per branch, as bars in the tree}
\usage{
plot_tree_contribution_bars(
  tree,
  signatures = NULL,
  contribution = NULL,
  mut_matrix = NULL,
  signature_colors = NULL,
  scaling = 1,
  remove_min = 20,
  ...
)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}

\item{signatures}{a matrix with signature names in the columns and mutations in the rows}

\item{contribution}{a contribution matrix with signatures in the rows and samples in the columns}

\item{mut_matrix}{a mutational matrix with samples in the columns and mutations in the rows}

\item{scaling}{how to scale the bars. If the bars are too big/small, this can be changed to adjust them}

\item{remove_min}{columns with equal or fewer mutations that this number will be removed}

\item{...}{parameters to pass on to \link{plot_gg_tree}}
}
\value{
a ggplot
}
\description{
plot the tree with the contribution of signature per branch, as bars in the tree
}
