% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_tree}
\alias{prepare_tree}
\title{Adds branch_ids to a tree}
\usage{
prepare_tree(tree)
}
\arguments{
\item{tree}{a \code{\link[treeio]{treedata}} object created by "cellPhyWrapper"}
}
\value{
a tree, with a \code{branch_id} column for branch identification
}
\description{
Adds branch_ids to a tree
}
